+++ autd3-core/src/sampling_config/mod.rs
@@ -3,7 +3,7 @@ mod error;
 use std::{fmt::Debug, num::NonZeroU16};
 
 use crate::{
-    defined::{Freq, Hz, ultrasound_freq},
+    defined::{Freq, Hz, ULTRASOUND_FREQ},
     utils::float::is_integer,
 };
 
@@ -20,12 +20,10 @@ pub enum SamplingConfig {
     Division(NonZeroU16),
     #[doc(hidden)]
     Freq(Freq<f32>),
-    #[cfg(not(feature = "dynamic_freq"))]
     #[doc(hidden)]
     Period(std::time::Duration),
     #[doc(hidden)]
     FreqNearest(Nearest<Freq<f32>>),
-    #[cfg(not(feature = "dynamic_freq"))]
     #[doc(hidden)]
     PeriodNearest(Nearest<std::time::Duration>),
 }
@@ -44,12 +42,10 @@ impl std::fmt::Debug for SamplingConfig {
         match self {
             SamplingConfig::Division(div) => write!(f, "SamplingConfig::Division({})", div),
             SamplingConfig::Freq(freq) => write!(f, "SamplingConfig::Freq({:?})", freq),
-            #[cfg(not(feature = "dynamic_freq"))]
             SamplingConfig::Period(period) => write!(f, "SamplingConfig::Period({:?})", period),
             SamplingConfig::FreqNearest(nearest) => {
                 write!(f, "SamplingConfig::FreqNearest({:?})", nearest)
             }
-            #[cfg(not(feature = "dynamic_freq"))]
             SamplingConfig::PeriodNearest(nearest) => {
                 write!(f, "SamplingConfig::PeriodNearest({:?})", nearest)
             }
@@ -69,7 +65,6 @@ impl From<Freq<f32>> for SamplingConfig {
     }
 }
 
-#[cfg(not(feature = "dynamic_freq"))]
 impl From<std::time::Duration> for SamplingConfig {
     fn from(value: std::time::Duration) -> Self {
         Self::Period(value)
@@ -90,52 +85,50 @@ impl SamplingConfig {
 
     /// The division number of the sampling frequency.
     ///
-    /// The sampling frequency is [`ultrasound_freq`] / `division`.
+    /// The sampling frequency is [`ULTRASOUND_FREQ`] / `division`.
     pub fn division(&self) -> Result<u16, SamplingConfigError> {
         match *self {
             SamplingConfig::Division(div) => Ok(div.get()),
             SamplingConfig::Freq(freq) => {
-                let freq_max = ultrasound_freq().hz() as f32 * Hz;
+                let freq_max = ULTRASOUND_FREQ.hz() as f32 * Hz;
                 let freq_min = freq_max / u16::MAX as f32;
                 if !(freq_min..=freq_max).contains(&freq) {
                     return Err(SamplingConfigError::FreqOutOfRangeF(
                         freq, freq_min, freq_max,
                     ));
                 }
-                let division = ultrasound_freq().hz() as f32 / freq.hz();
+                let division = ULTRASOUND_FREQ.hz() as f32 / freq.hz();
                 if !is_integer(division as _) {
                     return Err(SamplingConfigError::FreqInvalidF(freq));
                 }
                 Ok(division as _)
             }
-            #[cfg(not(feature = "dynamic_freq"))]
             SamplingConfig::Period(duration) => {
-                use crate::defined::ultrasound_period;
+                use crate::defined::ULTRASOUND_PERIOD;
 
-                let period_min = ultrasound_period();
+                let period_min = ULTRASOUND_PERIOD;
                 let period_max = std::time::Duration::from_micros(
-                    u16::MAX as u64 * ultrasound_period().as_micros() as u64,
+                    u16::MAX as u64 * ULTRASOUND_PERIOD.as_micros() as u64,
                 );
                 if !(period_min..=period_max).contains(&duration) {
                     return Err(SamplingConfigError::PeriodOutOfRange(
                         duration, period_min, period_max,
                     ));
                 }
-                if duration.as_nanos() % ultrasound_period().as_nanos() != 0 {
+                if duration.as_nanos() % ULTRASOUND_PERIOD.as_nanos() != 0 {
                     return Err(SamplingConfigError::PeriodInvalid(duration));
                 }
-                Ok((duration.as_nanos() / ultrasound_period().as_nanos()) as _)
+                Ok((duration.as_nanos() / ULTRASOUND_PERIOD.as_nanos()) as _)
             }
-            SamplingConfig::FreqNearest(nearest) => Ok((ultrasound_freq().hz() as f32
+            SamplingConfig::FreqNearest(nearest) => Ok((ULTRASOUND_FREQ.hz() as f32
                 / nearest.0.hz())
             .clamp(1.0, u16::MAX as f32)
             .round() as u16),
-            #[cfg(not(feature = "dynamic_freq"))]
             SamplingConfig::PeriodNearest(nearest) => {
-                use crate::defined::ultrasound_period;
+                use crate::defined::ULTRASOUND_PERIOD;
 
-                Ok(((nearest.0.as_nanos() + ultrasound_period().as_nanos() / 2)
-                    / ultrasound_period().as_nanos())
+                Ok(((nearest.0.as_nanos() + ULTRASOUND_PERIOD.as_nanos() / 2)
+                    / ULTRASOUND_PERIOD.as_nanos())
                 .clamp(1, u16::MAX as u128) as u16)
             }
         }
@@ -143,13 +136,12 @@ impl SamplingConfig {
 
     /// The sampling frequency.
     pub fn freq(&self) -> Result<Freq<f32>, SamplingConfigError> {
-        Ok(ultrasound_freq().hz() as f32 / self.division()? as f32 * Hz)
+        Ok(ULTRASOUND_FREQ.hz() as f32 / self.division()? as f32 * Hz)
     }
 
-    #[cfg(not(feature = "dynamic_freq"))]
     /// The sampling period.
     pub fn period(&self) -> Result<std::time::Duration, SamplingConfigError> {
-        Ok(crate::defined::ultrasound_period() * self.division()? as u32)
+        Ok(crate::defined::ULTRASOUND_PERIOD * self.division()? as u32)
     }
 }
 
@@ -159,7 +151,6 @@ impl SamplingConfig {
     pub const fn into_nearest(self) -> SamplingConfig {
         match self {
             SamplingConfig::Freq(freq) => SamplingConfig::FreqNearest(Nearest(freq)),
-            #[cfg(not(feature = "dynamic_freq"))]
             SamplingConfig::Period(period) => SamplingConfig::PeriodNearest(Nearest(period)),
             _ => self,
         }
@@ -170,9 +161,7 @@ impl SamplingConfig {
 mod tests {
     use crate::defined::{Hz, kHz};
 
-    #[cfg(not(feature = "dynamic_freq"))]
-    use crate::defined::ultrasound_period;
-    #[cfg(not(feature = "dynamic_freq"))]
+    use crate::defined::ULTRASOUND_PERIOD;
     use std::time::Duration;
 
     use super::*;
@@ -183,17 +172,14 @@ mod tests {
     #[case(Ok(u16::MAX), NonZeroU16::MAX)]
     #[case(Ok(1), 40000. * Hz)]
     #[case(Ok(10), 4000. * Hz)]
-    #[case(Err(SamplingConfigError::FreqInvalidF((ultrasound_freq().hz() as f32 - 1.) * Hz)), (ultrasound_freq().hz() as f32 - 1.) * Hz)]
-    #[case(Err(SamplingConfigError::FreqOutOfRangeF(0. * Hz, ultrasound_freq().hz() as f32 * Hz / u16::MAX as f32, ultrasound_freq().hz() as f32 * Hz)), 0. * Hz)]
-    #[case(Err(SamplingConfigError::FreqOutOfRangeF(40000. * Hz + 1. * Hz, ultrasound_freq().hz() as f32 * Hz / u16::MAX as f32, ultrasound_freq().hz() as f32 * Hz)), 40000. * Hz + 1. * Hz)]
-    #[cfg_attr(not(feature = "dynamic_freq"), case(Ok(1), Duration::from_micros(25)))]
-    #[cfg_attr(
-        not(feature = "dynamic_freq"),
-        case(Ok(10), Duration::from_micros(250))
-    )]
-    #[cfg_attr(not(feature = "dynamic_freq"), case(Err(SamplingConfigError::PeriodInvalid(Duration::from_micros(u16::MAX as u64 * ultrasound_period().as_micros() as u64) - Duration::from_nanos(1))), Duration::from_micros(u16::MAX as u64 * ultrasound_period().as_micros() as u64) - Duration::from_nanos(1)))]
-    #[cfg_attr(not(feature = "dynamic_freq"), case(Err(SamplingConfigError::PeriodOutOfRange(ultrasound_period() / 2, ultrasound_period(), Duration::from_micros(u16::MAX as u64 * ultrasound_period().as_micros() as u64))), ultrasound_period() / 2))]
-    #[cfg_attr(not(feature = "dynamic_freq"), case(Err(SamplingConfigError::PeriodOutOfRange(Duration::from_micros(u16::MAX as u64 * ultrasound_period().as_micros() as u64) * 2, ultrasound_period(), Duration::from_micros(u16::MAX as u64 * ultrasound_period().as_micros() as u64))), Duration::from_micros(u16::MAX as u64 * ultrasound_period().as_micros() as u64) * 2))]
+    #[case(Err(SamplingConfigError::FreqInvalidF((ULTRASOUND_FREQ.hz() as f32 - 1.) * Hz)), (ULTRASOUND_FREQ.hz() as f32 - 1.) * Hz)]
+    #[case(Err(SamplingConfigError::FreqOutOfRangeF(0. * Hz, ULTRASOUND_FREQ.hz() as f32 * Hz / u16::MAX as f32, ULTRASOUND_FREQ.hz() as f32 * Hz)), 0. * Hz)]
+    #[case(Err(SamplingConfigError::FreqOutOfRangeF(40000. * Hz + 1. * Hz, ULTRASOUND_FREQ.hz() as f32 * Hz / u16::MAX as f32, ULTRASOUND_FREQ.hz() as f32 * Hz)), 40000. * Hz + 1. * Hz)]
+    #[case(Ok(1), Duration::from_micros(25))]
+    #[case(Ok(10), Duration::from_micros(250))]
+    #[case(Err(SamplingConfigError::PeriodInvalid(Duration::from_micros(u16::MAX as u64 * ULTRASOUND_PERIOD.as_micros() as u64) - Duration::from_nanos(1))), Duration::from_micros(u16::MAX as u64 * ULTRASOUND_PERIOD.as_micros() as u64) - Duration::from_nanos(1))]
+    #[case(Err(SamplingConfigError::PeriodOutOfRange(ULTRASOUND_PERIOD / 2, ULTRASOUND_PERIOD, Duration::from_micros(u16::MAX as u64 * ULTRASOUND_PERIOD.as_micros() as u64))), ULTRASOUND_PERIOD / 2)]
+    #[case(Err(SamplingConfigError::PeriodOutOfRange(Duration::from_micros(u16::MAX as u64 * ULTRASOUND_PERIOD.as_micros() as u64) * 2, ULTRASOUND_PERIOD, Duration::from_micros(u16::MAX as u64 * ULTRASOUND_PERIOD.as_micros() as u64))), Duration::from_micros(u16::MAX as u64 * ULTRASOUND_PERIOD.as_micros() as u64) * 2)]
     fn division(
         #[case] expect: Result<u16, SamplingConfigError>,
         #[case] value: impl Into<SamplingConfig>,
@@ -207,8 +193,8 @@ mod tests {
     #[case(Ok(0.61036086 * Hz), NonZeroU16::MAX)]
     #[case(Ok(40000. * Hz), 40000. * Hz)]
     #[case(Ok(4000. * Hz), 4000. * Hz)]
-    #[cfg_attr(not(feature = "dynamic_freq"), case(Ok(40000. * Hz), Duration::from_micros(25)))]
-    #[cfg_attr(not(feature = "dynamic_freq"), case(Ok(4000. * Hz), Duration::from_micros(250)))]
+    #[case(Ok(40000. * Hz), Duration::from_micros(25))]
+    #[case(Ok(4000. * Hz), Duration::from_micros(250))]
     fn freq(
         #[case] expect: Result<Freq<f32>, SamplingConfigError>,
         #[case] value: impl Into<SamplingConfig>,
@@ -216,7 +202,6 @@ mod tests {
         assert_eq!(expect, SamplingConfig::new(value).freq());
     }
 
-    #[cfg(not(feature = "dynamic_freq"))]
     #[rstest::rstest]
     #[test]
     #[case(Ok(Duration::from_micros(25)), NonZeroU16::MIN)]
@@ -236,7 +221,7 @@ mod tests {
     #[test]
     #[case::min(u16::MAX, (40000. / u16::MAX as f32) * Hz)]
     #[case::max(1, 40000. * Hz)]
-    #[case::not_supported_max(1, (ultrasound_freq().hz() as f32 - 1.) * Hz)]
+    #[case::not_supported_max(1, (ULTRASOUND_FREQ.hz() as f32 - 1.) * Hz)]
     #[case::out_of_range_min(u16::MAX, 0. * Hz)]
     #[case::out_of_range_max(1, 40000. * Hz + 1. * Hz)]
     fn from_freq_nearest(#[case] expected: u16, #[case] freq: Freq<f32>) {
@@ -246,14 +231,13 @@ mod tests {
         );
     }
 
-    #[cfg(not(feature = "dynamic_freq"))]
     #[rstest::rstest]
     #[test]
-    #[case::min(1, ultrasound_period())]
-    #[case::max(u16::MAX, Duration::from_micros(u16::MAX as u64 * ultrasound_period().as_micros() as u64))]
-    #[case::not_supported_max(u16::MAX, Duration::from_micros(u16::MAX as u64 * ultrasound_period().as_micros() as u64) - Duration::from_nanos(1))]
-    #[case::out_of_range_min(1, ultrasound_period() / 2)]
-    #[case::out_of_range_max(u16::MAX, Duration::from_micros(u16::MAX as u64 * ultrasound_period().as_micros() as u64) * 2)]
+    #[case::min(1, ULTRASOUND_PERIOD)]
+    #[case::max(u16::MAX, Duration::from_micros(u16::MAX as u64 * ULTRASOUND_PERIOD.as_micros() as u64))]
+    #[case::not_supported_max(u16::MAX, Duration::from_micros(u16::MAX as u64 * ULTRASOUND_PERIOD.as_micros() as u64) - Duration::from_nanos(1))]
+    #[case::out_of_range_min(1, ULTRASOUND_PERIOD / 2)]
+    #[case::out_of_range_max(u16::MAX, Duration::from_micros(u16::MAX as u64 * ULTRASOUND_PERIOD.as_micros() as u64) * 2)]
     fn from_period_nearest(#[case] expected: u16, #[case] p: Duration) {
         assert_eq!(
             Ok(expected),
@@ -267,20 +251,14 @@ mod tests {
         SamplingConfig::Division(NonZeroU16::MIN)
     )]
     #[case(SamplingConfig::FreqNearest(Nearest(1. * Hz)), SamplingConfig::Freq(1. * Hz))]
-    #[cfg_attr(
-        not(feature = "dynamic_freq"),
-        case(
-            SamplingConfig::PeriodNearest(Nearest(Duration::from_micros(1))),
-            SamplingConfig::Period(Duration::from_micros(1))
-        )
+    #[case(
+        SamplingConfig::PeriodNearest(Nearest(Duration::from_micros(1))),
+        SamplingConfig::Period(Duration::from_micros(1))
     )]
     #[case(SamplingConfig::FreqNearest(Nearest(1. * Hz)), SamplingConfig::FreqNearest(Nearest(1. * Hz)))]
-    #[cfg_attr(
-        not(feature = "dynamic_freq"),
-        case(
-            SamplingConfig::PeriodNearest(Nearest(Duration::from_micros(1))),
-            SamplingConfig::PeriodNearest(Nearest(Duration::from_micros(1)))
-        )
+    #[case(
+        SamplingConfig::PeriodNearest(Nearest(Duration::from_micros(1))),
+        SamplingConfig::PeriodNearest(Nearest(Duration::from_micros(1)))
     )]
     #[test]
     fn into_nearest(#[case] expect: SamplingConfig, #[case] config: SamplingConfig) {
@@ -291,13 +269,10 @@ mod tests {
     #[case(true, SamplingConfig::FREQ_40K, SamplingConfig::FREQ_40K)]
     #[case(true, SamplingConfig::FREQ_40K, SamplingConfig::new(NonZeroU16::MIN))]
     #[case(true, SamplingConfig::FREQ_40K, SamplingConfig::new(40. * kHz))]
-    #[cfg_attr(
-        not(feature = "dynamic_freq"),
-        case(
-            true,
-            SamplingConfig::FREQ_40K,
-            SamplingConfig::new(std::time::Duration::from_micros(25))
-        )
+    #[case(
+        true,
+        SamplingConfig::FREQ_40K,
+        SamplingConfig::new(std::time::Duration::from_micros(25))
     )]
     #[case(false, SamplingConfig::new(41. * kHz), SamplingConfig::new(41. * kHz))]
     #[test]
@@ -311,20 +286,14 @@ mod tests {
         SamplingConfig::Division(NonZeroU16::MIN)
     )]
     #[case("SamplingConfig::Freq(1 Hz)", SamplingConfig::Freq(1. * Hz))]
-    #[cfg_attr(
-        not(feature = "dynamic_freq"),
-        case(
-            "SamplingConfig::Period(1µs)",
-            SamplingConfig::Period(Duration::from_micros(1))
-        )
+    #[case(
+        "SamplingConfig::Period(1µs)",
+        SamplingConfig::Period(Duration::from_micros(1))
     )]
     #[case("SamplingConfig::FreqNearest(Nearest(1 Hz))", SamplingConfig::FreqNearest(Nearest(1. * Hz)))]
-    #[cfg_attr(
-        not(feature = "dynamic_freq"),
-        case(
-            "SamplingConfig::PeriodNearest(Nearest(1µs))",
-            SamplingConfig::PeriodNearest(Nearest(Duration::from_micros(1)))
-        )
+    #[case(
+        "SamplingConfig::PeriodNearest(Nearest(1µs))",
+        SamplingConfig::PeriodNearest(Nearest(Duration::from_micros(1)))
     )]
     #[test]
     fn debug(#[case] expect: &str, #[case] config: SamplingConfig) {