+++ autd3-driver/src/firmware/operation/pulse_width_encoder.rs
@@ -2,13 +2,13 @@ use std::{convert::Infallible, mem::size_of};
 
 use crate::{
     firmware::{
-        fpga::PWE_BUF_SIZE,
+        fpga::{PWE_BUF_SIZE, PulseWidth},
         operation::{Operation, TypeTag},
     },
     geometry::Device,
 };
 
-use autd3_core::gain::EmitIntensity;
+use autd3_core::{defined::ULTRASOUND_PERIOD_COUNT, gain::EmitIntensity};
 use zerocopy::{Immutable, IntoBytes};
 
 #[repr(C, align(2))]
@@ -18,18 +18,20 @@ struct Pwe {
     __: u8,
 }
 
-pub struct PulseWidthEncoderOp<F: Fn(EmitIntensity) -> u8> {
+pub struct PulseWidthEncoderOp<F: Fn(EmitIntensity) -> PulseWidth<ULTRASOUND_PERIOD_COUNT>> {
     is_done: bool,
     f: F,
 }
 
-impl<F: Fn(EmitIntensity) -> u8> PulseWidthEncoderOp<F> {
+impl<F: Fn(EmitIntensity) -> PulseWidth<ULTRASOUND_PERIOD_COUNT>> PulseWidthEncoderOp<F> {
     pub(crate) const fn new(f: F) -> Self {
         Self { is_done: false, f }
     }
 }
 
-impl<F: Fn(EmitIntensity) -> u8 + Send + Sync> Operation for PulseWidthEncoderOp<F> {
+impl<F: Fn(EmitIntensity) -> PulseWidth<ULTRASOUND_PERIOD_COUNT> + Send + Sync> Operation
+    for PulseWidthEncoderOp<F>
+{
     type Error = Infallible;
 
     fn pack(&mut self, _: &Device, tx: &mut [u8]) -> Result<usize, Self::Error> {
@@ -42,11 +44,11 @@ impl<F: Fn(EmitIntensity) -> u8 + Send + Sync> Operation for PulseWidthEncoderOp
         );
 
         tx[size_of::<Pwe>()..]
-            .iter_mut()
+            .chunks_mut(size_of::<u16>())
             .take(PWE_BUF_SIZE)
             .enumerate()
-            .for_each(|(i, x)| {
-                *x = (self.f)(EmitIntensity(i as u8));
+            .for_each(|(i, dst)| {
+                super::write_to_tx(dst, (self.f)(EmitIntensity(i as u8)).0);
             });
 
         self.is_done = true;
@@ -76,9 +78,9 @@ mod tests {
     fn test() {
         let device = create_device(NUM_TRANS_IN_UNIT);
 
-        let mut tx = [0x00u8; 2 * (size_of::<Pwe>() + PWE_BUF_SIZE)];
+        let mut tx = [0x00u8; 2 * (size_of::<Pwe>() + PWE_BUF_SIZE * size_of::<u16>())];
 
-        let mut op = PulseWidthEncoderOp::new(|i| i.0);
+        let mut op = PulseWidthEncoderOp::new(|i| PulseWidth(i.0 as u16));
 
         assert_eq!(size_of::<Pwe>() + PWE_BUF_SIZE, op.required_size(&device));
 
@@ -89,6 +91,6 @@ mod tests {
         assert!(op.is_done());
 
         assert_eq!(tx[0], TypeTag::ConfigPulseWidthEncoder as u8);
-        assert!((0..PWE_BUF_SIZE).all(|i| i as u8 == tx[2 + i]));
+        assert!((0..PWE_BUF_SIZE).all(|i| i as u8 == tx[2 + 2 * i]));
     }
 }