+++ autd3-firmware-emulator/src/cpu/operation/stm/gain.rs
@@ -80,27 +80,21 @@ impl CPUEmulator {
                 self.stm_transition_value = d.head.transition_value;
                 self.stm_freq_div[segment as usize] = d.head.freq_div;
 
-                match segment {
-                    0 => {
-                        self.bram_write(
-                            BRAM_SELECT_CONTROLLER,
-                            ADDR_STM_FREQ_DIV0,
-                            d.head.freq_div,
-                        );
-                        self.bram_write(BRAM_SELECT_CONTROLLER, ADDR_STM_MODE0, STM_MODE_GAIN);
-                        self.bram_write(BRAM_SELECT_CONTROLLER, ADDR_STM_REP0, d.head.rep);
-                    }
-                    1 => {
-                        self.bram_write(
-                            BRAM_SELECT_CONTROLLER,
-                            ADDR_STM_FREQ_DIV1,
-                            d.head.freq_div,
-                        );
-                        self.bram_write(BRAM_SELECT_CONTROLLER, ADDR_STM_MODE1, STM_MODE_GAIN);
-                        self.bram_write(BRAM_SELECT_CONTROLLER, ADDR_STM_REP1, d.head.rep);
-                    }
-                    _ => unreachable!(),
-                }
+                self.bram_write(
+                    BRAM_SELECT_CONTROLLER,
+                    ADDR_STM_FREQ_DIV0 + segment as u16,
+                    d.head.freq_div,
+                );
+                self.bram_write(
+                    BRAM_SELECT_CONTROLLER,
+                    ADDR_STM_MODE0 + segment as u16,
+                    STM_MODE_GAIN,
+                );
+                self.bram_write(
+                    BRAM_SELECT_CONTROLLER,
+                    ADDR_STM_REP0 + segment as u16,
+                    d.head.rep,
+                );
 
                 self.change_stm_wr_segment(segment as _);
                 self.change_stm_wr_page(0);
@@ -111,7 +105,8 @@ impl CPUEmulator {
             };
 
             let mut src = src_base;
-            let mut dst = (self.stm_cycle[segment as usize] & GAIN_STM_BUF_PAGE_SIZE_MASK) << 8;
+            let mut dst =
+                ((self.stm_cycle[segment as usize] as u16) & GAIN_STM_BUF_PAGE_SIZE_MASK) << 8;
 
             match self.gain_stm_mode {
                 GAIN_STM_MODE_INTENSITY_PHASE_FULL => {
@@ -132,8 +127,9 @@ impl CPUEmulator {
 
                     if send > 1 {
                         let mut src = src_base;
-                        let mut dst =
-                            (self.stm_cycle[segment as usize] & GAIN_STM_BUF_PAGE_SIZE_MASK) << 8;
+                        let mut dst = ((self.stm_cycle[segment as usize] as u16)
+                            & GAIN_STM_BUF_PAGE_SIZE_MASK)
+                            << 8;
                         (0..self.num_transducers).for_each(|_| {
                             self.bram_write(
                                 BRAM_SELECT_STM,
@@ -157,8 +153,9 @@ impl CPUEmulator {
 
                     if send > 1 {
                         let mut src = src_base;
-                        let mut dst =
-                            (self.stm_cycle[segment as usize] & GAIN_STM_BUF_PAGE_SIZE_MASK) << 8;
+                        let mut dst = ((self.stm_cycle[segment as usize] as u16)
+                            & GAIN_STM_BUF_PAGE_SIZE_MASK)
+                            << 8;
                         (0..self.num_transducers).for_each(|_| {
                             let phase = (src.read() >> 4) & 0x000F;
                             self.bram_write(BRAM_SELECT_STM, dst, 0xFF00 | (phase << 4) | phase);
@@ -170,8 +167,9 @@ impl CPUEmulator {
 
                     if send > 2 {
                         let mut src = src_base;
-                        let mut dst =
-                            (self.stm_cycle[segment as usize] & GAIN_STM_BUF_PAGE_SIZE_MASK) << 8;
+                        let mut dst = ((self.stm_cycle[segment as usize] as u16)
+                            & GAIN_STM_BUF_PAGE_SIZE_MASK)
+                            << 8;
                         (0..self.num_transducers).for_each(|_| {
                             let phase = (src.read() >> 8) & 0x000F;
                             self.bram_write(BRAM_SELECT_STM, dst, 0xFF00 | (phase << 4) | phase);
@@ -183,8 +181,9 @@ impl CPUEmulator {
 
                     if send > 3 {
                         let mut src = src_base;
-                        let mut dst =
-                            (self.stm_cycle[segment as usize] & GAIN_STM_BUF_PAGE_SIZE_MASK) << 8;
+                        let mut dst = ((self.stm_cycle[segment as usize] as u16)
+                            & GAIN_STM_BUF_PAGE_SIZE_MASK)
+                            << 8;
                         (0..self.num_transducers).for_each(|_| {
                             let phase = (src.read() >> 12) & 0x000F;
                             self.bram_write(BRAM_SELECT_STM, dst, 0xFF00 | (phase << 4) | phase);
@@ -197,32 +196,21 @@ impl CPUEmulator {
                 _ => return ERR_INVALID_GAIN_STM_MODE,
             }
 
-            if self.stm_cycle[segment as usize] & GAIN_STM_BUF_PAGE_SIZE_MASK == 0 {
+            if (self.stm_cycle[segment as usize] as u16) & GAIN_STM_BUF_PAGE_SIZE_MASK == 0 {
                 self.change_stm_wr_page(
-                    ((self.stm_cycle[segment as usize] & !GAIN_STM_BUF_PAGE_SIZE_MASK)
+                    (((self.stm_cycle[segment as usize] as u16) & !GAIN_STM_BUF_PAGE_SIZE_MASK)
                         >> GAIN_STM_BUF_PAGE_SIZE_WIDTH) as _,
                 );
             }
 
             if (d.subseq.flag & GAIN_STM_FLAG_END) == GAIN_STM_FLAG_END {
                 self.stm_mode[segment as usize] = STM_MODE_GAIN;
-                match segment {
-                    0 => {
-                        self.bram_write(
-                            BRAM_SELECT_CONTROLLER,
-                            ADDR_STM_CYCLE0,
-                            (self.stm_cycle[segment as usize].max(1) - 1) as _,
-                        );
-                    }
-                    1 => {
-                        self.bram_write(
-                            BRAM_SELECT_CONTROLLER,
-                            ADDR_STM_CYCLE1,
-                            (self.stm_cycle[segment as usize].max(1) - 1) as _,
-                        );
-                    }
-                    _ => unreachable!(),
-                }
+
+                self.bram_write(
+                    BRAM_SELECT_CONTROLLER,
+                    ADDR_STM_CYCLE0 + segment as u16,
+                    (self.stm_cycle[segment as usize].max(1) - 1) as _,
+                );
 
                 if (d.subseq.flag & GAIN_STM_FLAG_UPDATE) == GAIN_STM_FLAG_UPDATE {
                     return self.stm_segment_update(