+++ autd3-firmware-emulator/tests/op/stm/gain.rs
@@ -12,7 +12,7 @@ use autd3_driver::{
         fpga::{
             Drive, EmitIntensity, GAIN_STM_BUF_SIZE_MAX, GPIOIn, LoopBehavior, Phase,
             SILENCER_STEPS_INTENSITY_DEFAULT, SILENCER_STEPS_PHASE_DEFAULT, SamplingConfig,
-            Segment, SilencerTarget, TransitionMode,
+            Segment, TransitionMode,
         },
         operation::OperationHandler,
     },
@@ -344,7 +344,6 @@ fn gain_stm_freq_div_too_small() -> anyhow::Result<()> {
                 phase: NonZeroU16::new(SILENCER_STEPS_PHASE_DEFAULT * 2).unwrap(),
                 strict_mode: true,
             },
-            target: SilencerTarget::Intensity,
         };
         assert_eq!(Ok(()), send(&mut cpu, d, &geometry, &mut tx));